Houston, we have a problem.
Either the server is crushing race data, or is not calculating correctly.

Becomes more noticeable with time, but for example:
Driver 1 is on Lap 10, Sector 3 but has a higher total time than Driver 2 on Lap 9, Sector 1

Otherwise, the results are now being ordered by Lap, Sector, then Time. Its fun to watch ;)

Time of sleep of pilots has been extended, so its a bit slower.